FEAT: Added mask functionality to the multi-statistics filter in input #1. (Data goes in input #0.)

+++ vtkRobartsUtilities/vtkImageMultiStatistics.h
@@ -57,9 +57,9 @@ protected:
 
   void Execute();
   
-  double* AverageMagnitude;
-  double* Covariance;
-  double* JointEntropy;
+  long double* AverageMagnitude;
+  long double* Covariance;
+  long double* JointEntropy;
   long int Count;
   int NumberOfComponents;
   int NumberOfBins;
